Change to int type of EntityId

+++ src/ACs.NHibernate.Generic/IEntityId.cs
@@ -2,6 +2,7 @@
 {
     public interface IEntityId
     {
-        long Id { get; }
+        int Id { get; }
     }
+
 }
