+++ src/ACs.NHibernate.Generic/IRepository.cs
@@ -3,7 +3,7 @@ namespace ACs.NHibernate.Generic
 {
     public interface IRepository<T> where T : IEntityRoot, IEntityId
     {
-        T GetById(long id);
+        T GetById(int id);
         void Save(T entity);
     }
 }
